(feat) move dummylist components into dedicated components and fix some linting errors
