+++ client/components/dummyList/index.js
@@ -1,0 +1,2 @@
+export DummyList from './DummyList';
+export DummyListItem from './DummyListItem';
